requirements.txt
numpy
torch
torchvision
torchaudio
gorilla-core
open_clip_torch
pytorchse3